// example-app/src/swagger-generated/accounts/api.ts
export interface Account {
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    accountId: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    productName: string;
    /**
     * 
     * @type {IdentifierSchema}
     * @memberof Account
     */
    identifierSchema: IdentifierSchema;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    identifier: string;
    /**
     * 
     * @type {ServicerSchema}
     * @memberof Account
     */
    servicerSchema: ServicerSchema;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    servicer: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    owner: string;
    /**
     * Monetary amount
     * @type {string}
     * @memberof Account
     */
    netBalance: string;
    /**
     * Monetary amount
     * @type {string}
     * @memberof Account
     */
    grossBalance: string;
    /**
     * Monetary amount
     * @type {string}
     * @memberof Account
     */
    counterBalance?: string;
    /**
     * Monetary amount
     * @type {string}
     * @memberof Account
     */
    coverReservationAmount?: string;
    /**
     * ISO 4127 currency code
     * @type {string}
     * @memberof Account
     */
    currency: string;
}
// ...
export interface Card {
    /**
     * 
     * @type {string}
     * @memberof Card
     */
    cardId: string;
    /**
     * 
     * @type {string}
     * @memberof Card
     */
    productName: string;
    /**
     * Masked card number
     * @type {string}
     * @memberof Card
     */
    cardNumber: string;
}

/**
 * 
 * @export
 * @interface CardTransaction
 */
export interface CardTransaction {
    /**
     * 
     * @type {string}
     * @memberof CardTransaction
     */
    description?: string;
    /**
     * Monetary amount
     * @type {string}
     * @memberof CardTransaction
     */
    amount?: string;
    /**
     * ISO 4127 currency code
     * @type {string}
     * @memberof CardTransaction
     */
    currency?: string;
    /**
     * Monetary amount
     * @type {string}
     * @memberof CardTransaction
     */
    originalAmount?: string;
    /**
     * ISO 4127 currency code
     * @type {string}
     * @memberof CardTransaction
     */
    originalCurrency?: string;
    /**
     * 
     * @type {string}
     * @memberof CardTransaction
     */
    postingDate?: string;
}

/**
 * 
 * @export
 * @interface CardTransactionsResponse
 */
export interface CardTransactionsResponse {
    /**
     * 
     * @type {Array<CardTransaction>}
     * @memberof CardTransactionsResponse
     */
    transactions?: Array<CardTransaction>;
    /**
     * 
     * @type {string}
     * @memberof CardTransactionsResponse
     */
    continuationToken?: string;
    /**
     * 
     * @type {TransactionsResponseLinks}
     * @memberof CardTransactionsResponse
     */
    links?: TransactionsResponseLinks;
}
// ...
export enum IdentifierSchema {
    IBAN = 'IBAN'
}
// ...
export interface Link {
    /**
     * 
     * @type {string}
     * @memberof Link
     */
    href?: string;
}

/**
 * 
 * @export
 * @interface Payer
 */
export interface Payer {
    /**
     * 
     * @type {string}
     * @memberof Payer
     */
    name?: string;
}

/**
 * 
 * @export
 * @interface Recipient
 */
export interface Recipient {
    /**
     * Name of the recipient
     * @type {string}
     * @memberof Recipient
     */
    name?: string;
    /**
     * Recipient account number typically in IBAN format for SEPA payments
     * @type {string}
     * @memberof Recipient
     */
    identifier?: string;
    /**
     * Recipieint bank identifier typically BIC for SEPA payments
     * @type {string}
     * @memberof Recipient
     */
    servicer?: string;
    /**
     * Recipient address
     * @type {Array<string>}
     * @memberof Recipient
     */
    address?: Array<string>;
}
// ...
export enum ServicerSchema {
    BIC = 'BIC'
}

/**
 * 
 * @export
 * @interface Transaction
 */
export interface Transaction {
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    archiveId?: string;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    message?: string;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    reference?: string;
    /**
     * Monetary amount
     * @type {string}
     * @memberof Transaction
     */
    amount?: string;
    /**
     * ISO 4127 currency code
     * @type {string}
     * @memberof Transaction
     */
    currency?: string;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    bookingDate?: string;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    valueDate?: string;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    paymentDate?: string;
    /**
     * 
     * @type {Payer}
     * @memberof Transaction
     */
    payer?: Payer;
    /**
     * 
     * @type {Recipient}
     * @memberof Transaction
     */
    recipient?: Recipient;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    proprietaryTransactionDescription?: string;
}

/**
 * 
 * @export
 * @interface TransactionsResponse
 */
export interface TransactionsResponse {
    /**
     * 
     * @type {Array<Transaction>}
     * @memberof TransactionsResponse
     */
    transactions?: Array<Transaction>;
    /**
     * 
     * @type {string}
     * @memberof TransactionsResponse
     */
    continuationToken?: string;
    /**
     * 
     * @type {TransactionsResponseLinks}
     * @memberof TransactionsResponse
     */
    links?: TransactionsResponseLinks;
}

/**
 * 
 * @export
 * @interface TransactionsResponseLinks
 */
export interface TransactionsResponseLinks {
    /**
     * 
     * @type {Link}
     * @memberof TransactionsResponseLinks
     */
    self?: Link;
    /**
     * 
     * @type {Link}
     * @memberof TransactionsResponseLinks
     */
    next?: Link;
}